Adding django and react files to start with
The react files are for a basic tic-tac-to tutorial I did. The django stuff
is completely unused, as it was provided in the tutorial for the docker image along
with a React frontend that got replaced.

Regardless, this should serve as a good baseline to be edited in our first meeting.